requirements.txt
torch>=1.9.0
numpy>=1.19.2
pyyaml>=5.4.1
scipy>=1.7.0 
pandas>=1.3.0
matplotlib>=3.4.3
